requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
alembic==1.7.6
antlr4-python3-runtime==4.8
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
audioread==2.1.9
banal==1.0.6
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
colorlog==6.6.0
coverage==5.5
cycler==0.10.0
datasets==1.18.3
decorator==5.0.9
dill==0.3.4
filelock==3.4.2
fsspec==2021.8.1
future==0.18.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.40.0
h5py==3.4.0
huggingface-hub==0.4.0
hydra-colorlog==1.1.0
hydra-core==1.1.1
idna==3.3
importlib-metadata==4.10.1
iniconfig==1.1.1
joblib==1.0.1
kiwisolver==1.3.2
librosa==0.8.1
llvmlite==0.37.0
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.4.3
multidict==5.1.0
multiprocess==0.70.12.2
numba==0.54.0
numpy==1.20.3
oauthlib==3.1.1
omegaconf==2.1.1
packaging==21.3
pandas==1.3.3
pescador==2.1.0
Pillow==8.3.2
pluggy==1.0.0
pooch==1.5.1
protobuf==3.18.0
py==1.10.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.1
pyparsing==3.0.7
pypeln==0.4.7
pytest==6.2.5
pytest-cov==2.12.1
python-dateutil==2.8.2
pytorch-lightning==1.4.7
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.2.1
requests==2.27.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
SoundFile==0.10.3.post1
SQLAlchemy==1.4.31
stopit==1.1.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.2.0
timm==0.5.4
toml==0.10.2
torch==1.9.1
torch-tb-profiler==0.3.1
torchaudio==0.9.1
torchlibrosa==0.0.9
torchlightning==0.0.0
torchmetrics==0.5.1
torchvision==0.10.1
tqdm==4.62.2
typing-extensions==3.10.0.2
urllib3==1.26.8
Werkzeug==2.0.1
xxhash==2.0.2
yarl==1.6.3
zipp==3.7.0
